Accès

+++ app/Http/Controllers/AuthController.php
@@ -7,7 +7,7 @@ use App\Models\User;
 use App\Models\SystemAdministrador;
 use App\Models\SystemModulo;
 use App\Models\SystemModuloMenu;
-use App\Models\Cuenta;
+use App\Models\SystemUserRol;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\DB;
 
@@ -27,21 +27,15 @@ class AuthController extends Controller
             'password' => bcrypt($fields['password'])
         ]);
 
-        // $token = $user->createToken('myapptoken')->plainTextToken;
-        if($request['cuenta_id']) {
-            $cuenta = Cuenta::find($request['cuenta_id']);
-
-            CuentaUsuario::create([
-                'cuenta_id' => $request['cuenta_id'],
+        if($request['rol_id']) {
+            SystemUserRol::create([
                 'user_id' => $user->id,
-                'rol_id' => 2,
-                'propietario_id' => $cuenta->user_id,
+                'rol_id' => $request['rol_id'],
             ]);
         }
 
         $response = [
             'user' => $user,
-            // 'token' => $token
         ];
 
         return response($response, 201);
@@ -55,7 +49,8 @@ class AuthController extends Controller
         ]);
         
         $user = User::where('email', $fields['email'])->first();
-        $user->tokens()->delete();
+        
+        if($user) $user->tokens()->delete();
 
         if(!$user || !Hash::check($fields['password'], $user->password)) {
             
@@ -69,17 +64,23 @@ class AuthController extends Controller
         $system_modulos = [];
         $administrador = SystemAdministrador::where("user_id", $user->id)->first();
 
-        if($administrador) {
-            $system_modulos = SystemModulo::all();
-
-            foreach($system_modulos as $k => $row) {
-                
-                $system_modulos[$k]['menus'] = SystemModuloMenu::where("modulo_id", $row->id)->get();
-            }
-        }
+        if($administrador) $system_modulos = SystemModulo::with("menus")->get();
         else {
-        }
+            $tmp = DB::table("users_accesos_modulos")
+                        ->select("modulo_id")
+                        ->where("user_id",$user->id)
+                        ->get();
 
+            $cond = [];
+
+            if($tmp->count()) {
+                foreach($tmp as $row) {
+                    $cond[] = $row->modulo_id;
+                }
+            }
+
+            $system_modulos = SystemModulo::whereIn("id", $cond)->with("menus")->get();
+        }
 
         $response = [
             'user' => $user,
@@ -104,7 +105,6 @@ class AuthController extends Controller
         $validated = [];
 
         if($request['password_reset']) {
-
             $fields = $request->validate([
                 'name' => 'required',
                 'password' => 'required|string|confirmed',
@@ -116,51 +116,17 @@ class AuthController extends Controller
             ];
         }
         else {
-            $fields = $request->validate([
-                'name' => 'required',
-            ]);
-
-            $validated = [
-                'name' => $fields['name'],
-            ];
+            $fields = $request->validate(['name' => 'required']);
+            $validated = ['name' => $fields['name']];
         }
 
         $user->update($validated);
 
-
-
-        $cuenta_usuario = CuentaUsuario::where("user_id", $user->id);
-        $request['user_id'] = $user->id;
-        // dd($request);
-
-        $fields = $request->validate([
-            'cuenta_id' => 'required',
-            'user_id' => 'required',
-            'rol_id' => 'required',
-            'propietario_id' => 'required',
-        ]);
-
-        if($cuenta_usuario) {
-
-            $cuenta_usuario->update([
-                'cuenta_id' => $fields['cuenta_id'],
-                'user_id' => $fields['user_id'],
-                'rol_id' => $fields['rol_id'],
-                'propietario_id' => $fields['propietario_id'],
-            ]);
+        if($request['rol_id']) {
+            DB::table('users_roles')
+                ->where("user_id",  $user->id)
+                ->update(['rol_id' => $request['rol_id']]);
         }
-        else {
-
-            $cuenta_usuario = CuentaUsuario::create([
-                'cuenta_id' => $fields['cuenta_id'],
-                'user_id' => $fields['cuenta_id'],
-                'rol_id' => $fields['rol_id'],
-                'propietario_id' => $fields['propietario_id'],
-            ]);
-        }
-
-
-
 
         return response()->json([
             'message' => 'User updated'
